requirements.txt
flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.1.1
pydantic==2.12.2
google-generativeai==0.8.5
requests==2.32.5